requirements.txt
ultralytics>=11.0.0
opencv-python>=4.8.0
fastapi>=0.108.0
uvicorn>=0.27.0
python-multipart>=0.0.6
numpy>=1.24.0
pillow>=10.0.0
supabase>=2.0.0
websockets>=12.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.28.0 